UPBGE: Refactor parallax UI.
Previously the parallax settings was put in the texture
influence panel, but they aren't only influence.

To make the UI clearer the parallaw settings are put in
a new panel named "Parallax" in texture panel.

In the same time the option "Use Parallax UV" is optional
if the texture is making the parallax UV.