UPBGE: Disallow depth transparency in cube map render.
Previously depth transparency was enabled for cube maps but
this technic request to copy or use a off screen containing
the depth. But in case of cube map the off screen can not be
used because it's a cube map texture.

To disallow this case a new render enumeration named
RAS_CUBEMAP is added. When the current drawing mode is
RAS_CUBEMAP the depth transparency meshes are not rendered.