Avoid to add diffuse over specular when only specular
is checked + display specular when the material is not enlightened.

First fix: Previously, when we only checked specular but not diffuse, the
diffuse was nevertheless computed and "mixed" with specular. To fix this,
we just add if !(lamp->mode & LA_NO_DIFF) before shade_area_diff_texture
shader.

Second fix: Previously, If the material which receive light was not
enlightened, his basic color was vec4(0.0, 0.0, 0.0, 1.0). This black
color was multiplied with texture color (
texture_rgb_blend(mat, tex_rgb, *rgb, GPU_uniform(&one),
GPU_uniform(&mtex->colfac), MTEX_MUL, rgb); )

When we used MTEX_MUL, we multiplied the texture color by black. The
result was a black color. To fix that, we change of blend type and instead of
MTEX_MUL, I choosed MTEX_BLEND so we can see the speclular projected by
the area light even if the "receiver" material is not enlightened.